Controller/requirements.txt
absl-py==1.1.0
asttokens==2.0.5
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
executing==0.8.3
filelock==3.7.1
fonttools==4.33.3
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.0
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.0.post2206102148
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.1.2